pyproject.toml
[tool.poetry]
name = "detaMVC"
version = "0.5.0"
description = ""
authors = ["MBeebe <grow.food.everywhere@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
detamvc = "detamvc.main:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.4.1"}
fastapi = "^0.78.0"
Jinja2 = "^3.1.2"
deta = "^1.1.0"
odetam = "^1.2.0"
python-multipart = "^0.0.5"
python-dotenv = "^0.20.0"
mkdocs-material = "^8.5.7"
authlib = "^1.2.0"
itsdangerous = "^2.1.2"
httpx = "^0.23.1"
pyjwt = "^2.6.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.dev-dependencies]
pytest = ">=5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
